CameraModule.java 팩토리 패턴 적용 유연성, 클라이언트 코드 단순화, 캡슐화
